modificati algoritmi di sorting in ramo sequenziale per sopportare tipo long

+++ sequenziale/src/sort.h
@@ -10,21 +10,20 @@
  * Pivot = data[alto] (last element)
  * Returns final pivot position
  */
-int partition_lomuto(int *data, int basso, int alto);
+int partition_lomuto(long *data, int basso,int alto);
 
 /**
  * Hoare partition scheme: bidirectional convergence
  * Pivot = data[basso] (first element, or pre-selected)
  * Returns pivot position
  */
-int partition_hoare(int *data, int basso, int alto);
+int partition_hoare(long *data,int basso,int alto);
 
-void quick_sort(int *data,int basso,int alto);
+void quick_sort(long *data,int basso,int alto);
 
-void select_median_of_3(int *data, int basso, int alto);
-void median_quick_sort(int *data,int basso,int alto);
+void select_median_of_3(long *data, int basso, int alto);
+void median_quick_sort(long *data,int basso,int alto);
 
-void merge(int* data,int low,int mid,int high);
+void merge(long* data,size_t low,size_t mid,size_t high);
 
-
-void merge_sort(int* data,int left,int right);
+void merge_sort(long* data,size_t left,size_t right);
